Shows status of bridges across Niagara canals.
It would be neat to know if the the canal bridges are up or down when
driving to Niagara on the Lake for example.  That information is posted
on the web:
http://greatlakes-seaway.com/R2/jsp/R2.jsp?language=E&loc=NiaBrdgStatus.
jsp  But it would be great if there was another way to access it that
make more sense, ie with a phone or by twitter?